Decompose god classes
